Move the PaginatedResource to its own module
